fix: make the Supabase deployment actually work, and verify it end to end
Live against the real project now: 2,535 rows in two schemas, 695 plates in
Storage, all 13 routes rendering correctly in a real browser.

Fixes found by running it rather than reading it
- postgres.js max:1 DEADLOCKS. Every page issues its queries with Promise.all,
  and with prepare:false (required by pgbouncer transaction mode) postgres.js
  cannot pipeline them down a single socket, so the first query holds the only
  connection and the rest wait forever. It does not error — the page streams a
  response that never terminates and looks like a slow network. Measured:
  max:1 hung past 45s, max:5 returned the same four queries in 0.83s.
- the client is cached on globalThis, so Next's dev reloads do not leak a new
  pool per edit until the pooler starts refusing connections
- feature-rows.tsx still linked crops at /api/shot, a route this migration
  deleted. It sits in components/, so the earlier sweep of app/ and lib/ missed
  it and every "view proof" link would have 404'd
- COUNT(*) is bigint and postgres.js returns bigint as a string, so the catalog
  counts were arriving as "84" rather than 84. Rendering hid it; arithmetic
  would not have. Cast to int in SQL, matching what pdp-db.ts already did

Schema hardening, from the supabase-postgres-best-practices skill
- index features.page_id and feature_links.canonical_id. Postgres does not
  index foreign keys for you, and the feature matrix joins canonical_id for
  every row it draws
- add a catalog_read role with SELECT and nothing else, so the deployed site
  never holds superuser credentials and the database password can be rotated
  without touching Vercel. Verified through the transaction pooler: reads both
  schemas, INSERT refused with InsufficientPrivilege

Verification tooling
- scripts/apply_schema.py, scripts/verify_online.py, scripts/browser_check.py
- verify_online compares every published count against the local SQLite, asserts
  the two schemas stay disjoint apart from luzzpickleball.com, and checks the
  FTS5 -> tsvector port actually returns rows
- browser_check opens all 13 routes in Chromium: 200 alone proves nothing, so it
  counts images that truly decoded, console errors and failed requests. It
  ignores Next's _rsc prefetch aborts, which fire on every route and mean nothing

+++ web/lib/db.ts
@@ -104,12 +104,12 @@ export async function getStats() {
 export function getSites() {
   return all_<Site & { feature_count: number; page_count: number; annotated: number }>(`
     SELECT s.*,
-      (SELECT COUNT(*) FROM catalog.features f WHERE f.site_id = s.id) AS feature_count,
+      (SELECT COUNT(*) FROM catalog.features f WHERE f.site_id = s.id)::int AS feature_count,
       (SELECT COUNT(*) FROM catalog.pages p
-        WHERE p.site_id = s.id AND p.scrape_status='ok')               AS page_count,
+        WHERE p.site_id = s.id AND p.scrape_status='ok')::int          AS page_count,
       (SELECT COUNT(*) FROM catalog.annotations a
          JOIN catalog.features f2 ON f2.id = a.feature_id
-        WHERE f2.site_id = s.id AND a.match_method <> 'failed')        AS annotated
+        WHERE f2.site_id = s.id AND a.match_method <> 'failed')::int   AS annotated
     FROM catalog.sites s
     ORDER BY feature_count DESC`);
 }
@@ -139,7 +139,7 @@ export function getPages(domain?: string) {
   const params = domain ? [domain] : [];
   return all_<PageRow>(
     `SELECT p.*, s.domain,
-        (SELECT COUNT(*) FROM catalog.features f WHERE f.page_id = p.id) AS feature_count,
+        (SELECT COUNT(*) FROM catalog.features f WHERE f.page_id = p.id)::int AS feature_count,
         (SELECT a.screenshot_path FROM catalog.annotations a
           WHERE a.page_id = p.id AND a.screenshot_path IS NOT NULL LIMIT 1) AS shot
       FROM catalog.pages p JOIN catalog.sites s ON s.id = p.site_id
@@ -152,7 +152,7 @@ export function getPages(domain?: string) {
 export function getPage(id: number) {
   return one<PageRow>(
     `SELECT p.*, s.domain,
-        (SELECT COUNT(*) FROM catalog.features f WHERE f.page_id = p.id) AS feature_count,
+        (SELECT COUNT(*) FROM catalog.features f WHERE f.page_id = p.id)::int AS feature_count,
         (SELECT a.screenshot_path FROM catalog.annotations a
           WHERE a.page_id = p.id AND a.screenshot_path IS NOT NULL LIMIT 1) AS shot
       FROM catalog.pages p JOIN catalog.sites s ON s.id = p.site_id
